move elderly defiinition from snap.py to usda.py
